+++ demo/src/main/java/com/braintreepayments/api/demo/DemoFragment.java
@@ -8,27 +8,22 @@ import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
 
+import androidx.annotation.IdRes;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.annotation.VisibleForTesting;
 import androidx.fragment.app.Fragment;
 
-import com.braintreepayments.api.BrowserSwitchClient;
 import com.braintreepayments.api.BrowserSwitchException;
-import com.braintreepayments.api.BrowserSwitchListener;
 import com.braintreepayments.api.BrowserSwitchOptions;
 import com.braintreepayments.api.BrowserSwitchResult;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
-public class DemoFragment extends Fragment implements View.OnClickListener, BrowserSwitchListener {
+public class DemoFragment extends Fragment implements View.OnClickListener {
 
     private static final String TEST_KEY = "testKey";
     private static final String TEST_VALUE = "testValue";
-
-    @VisibleForTesting
-    BrowserSwitchClient browserSwitchClient = null;
 
     private TextView mBrowserSwitchStatusTextView;
     private TextView mSelectedColorTextView;
@@ -54,27 +49,13 @@ public class DemoFragment extends Fragment implements View.OnClickListener, Brow
     }
 
     @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        browserSwitchClient = new BrowserSwitchClient();
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        browserSwitchClient.deliverResult(getActivity(), this);
-    }
-
-    @Override
     public void onClick(View v) {
-        switch (v.getId()) {
-            case R.id.browser_switch:
-                startBrowserSwitch();
-                break;
-            case R.id.browser_switch_with_metadata:
-                JSONObject metadata = buildMetadataObject(TEST_KEY, TEST_VALUE);
-                startBrowserSwitchWithMetadata(metadata);
-                break;
+        @IdRes int viewId = v.getId();
+        if (viewId == R.id.browser_switch) {
+            startBrowserSwitch();
+        } else if (viewId == R.id.browser_switch_with_metadata) {
+            JSONObject metadata = buildMetadataObject();
+            startBrowserSwitchWithMetadata(metadata);
         }
     }
 
@@ -83,9 +64,10 @@ public class DemoFragment extends Fragment implements View.OnClickListener, Brow
         BrowserSwitchOptions browserSwitchOptions = new BrowserSwitchOptions()
                 .requestCode(1)
                 .url(url)
-                .returnUrlScheme("my-custom-url-scheme");
+                .returnUrlScheme(getReturnUrlScheme());
+
         try {
-            browserSwitchClient.start(getActivity(), browserSwitchOptions);
+            getDemoActivity().startBrowserSwitch(browserSwitchOptions);
         } catch (BrowserSwitchException e) {
             String statusText = "Browser Switch Error: " + e.getMessage();
             mBrowserSwitchStatusTextView.setText(statusText);
@@ -99,10 +81,12 @@ public class DemoFragment extends Fragment implements View.OnClickListener, Brow
                 .metadata(metadata)
                 .requestCode(1)
                 .url(url)
-                .returnUrlScheme("my-custom-url-scheme");
+                .returnUrlScheme(getReturnUrlScheme());
         try {
-            browserSwitchClient.start(getActivity(), browserSwitchOptions);
+            getDemoActivity().startBrowserSwitch(browserSwitchOptions);
         } catch (BrowserSwitchException e) {
+            String statusText = "Browser Switch Error: " + e.getMessage();
+            mBrowserSwitchStatusTextView.setText(statusText);
             e.printStackTrace();
         }
     }
@@ -110,20 +94,19 @@ public class DemoFragment extends Fragment implements View.OnClickListener, Brow
     private Uri buildBrowserSwitchUrl() {
         String url = "https://braintree.github.io/popup-bridge-example/" +
                 "this_launches_in_popup.html?popupBridgeReturnUrlPrefix=" +
-                "my-custom-url-scheme://";
+                getReturnUrlScheme() + "://";
         return Uri.parse(url);
     }
 
-    private JSONObject buildMetadataObject(String key, String value) {
+    private JSONObject buildMetadataObject() {
         try {
-            return new JSONObject().put(key, value);
+            return new JSONObject().put(DemoFragment.TEST_KEY, DemoFragment.TEST_VALUE);
         } catch (JSONException ignore) {
             // do nothing
         }
         return null;
     }
 
-    @Override
     public void onBrowserSwitchResult(BrowserSwitchResult result) {
         String resultText = null;
         String selectedColorText = "";
@@ -157,5 +140,13 @@ public class DemoFragment extends Fragment implements View.OnClickListener, Brow
             }
         }
         mMetadataTextView.setText(String.format("Metadata: %s", metadataOutput));
+    }
+
+    private DemoActivity getDemoActivity() {
+        return (DemoActivity) getActivity();
+    }
+
+    private String getReturnUrlScheme() {
+        return getDemoActivity().getReturnUrlScheme();
     }
 }